+++ Code/Framework/AzCore/Tests/AZTestShared/Math/MathTestHelpers.cpp
@@ -1,6 +1,7 @@
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
@@ -94,7 +95,7 @@ namespace AZ
         auto printElement = [&os, &mat](int64_t row, int64_t col) -> std::ostream&
         {
             const std::streamsize width = 10;
-            os << std::setw(width) << std::fixed << mat.GetElement(row, col);
+            os << std::setw(width) << std::fixed << mat.GetElement(static_cast<int32_t>(row), static_cast<int32_t>(col));
             return os;
         };
