+++ Gems/AWSMetrics/Code/Tests/MetricsQueueTest.cpp
@@ -1,6 +1,7 @@
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
@@ -123,7 +124,7 @@ namespace AWSMetrics
             queue.AddMetrics(metrics);
         }
 
-        int maxCapacity = queue[0].GetSizeInBytes() * NumTestMetrics / 2;
+        int maxCapacity = static_cast<int>(queue[0].GetSizeInBytes() * NumTestMetrics / 2);
 
         ASSERT_EQ(queue.FilterMetricsByPriority(maxCapacity), NumTestMetrics / 2);
         ASSERT_EQ(queue.GetNumMetrics(), NumTestMetrics / 2);
@@ -229,7 +230,7 @@ namespace AWSMetrics
     {
         MetricsEvent metrics;
         metrics.AddAttribute(MetricsAttribute(AttrName, AttrValue));
-        int sizeOfEachMetrics = metrics.GetSizeInBytes();
+        int sizeOfEachMetrics = static_cast<int>(metrics.GetSizeInBytes());
 
         MetricsQueue queue;
         queue.AddMetrics(metrics);
