+++ Gems/AssetMemoryAnalyzer/Code/Source/ExportJSON.cpp
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
-#include "AssetMemoryAnalyzer_precompiled.h"
 
 #include "ExportJSON.h"
 
@@ -78,7 +78,7 @@ namespace AssetMemoryAnalyzer
             writer.StartObject();
 
             writer.Key("id");
-            writer.Int(idCounter++);
+            writer.Int(static_cast<int>(idCounter++));
 
             writer.Key("label");
             writer.String(asset.m_id ? asset.m_id : "Root");
@@ -98,7 +98,7 @@ namespace AssetMemoryAnalyzer
                 {
                     writer.StartObject();
                     writer.Key("id");
-                    writer.Int(idCounter++);
+                    writer.Int(static_cast<int>(idCounter++));
 
                     writer.Key("label");
                     writer.String("<local allocations>");
@@ -119,7 +119,7 @@ namespace AssetMemoryAnalyzer
 
                         writer.StartObject();
                         writer.Key("id");
-                        writer.Int(idCounter++);
+                        writer.Int(static_cast<int>(idCounter++));
 
                         writer.Key("label");
 
@@ -127,13 +127,13 @@ namespace AssetMemoryAnalyzer
                         {
                         case AllocationCategories::HEAP:
                             writer.String(FormatUtils::FormatCodePoint(*ap.m_codePoint));
-                            heapSummary.m_allocationCount = ap.m_allocations.size();
+                            heapSummary.m_allocationCount = static_cast<uint32_t>(ap.m_allocations.size());
                             heapSummary.m_allocatedMemory = ap.m_totalAllocatedMemory;
                             break;
 
                         case AllocationCategories::VRAM:
                             writer.String(ap.m_codePoint->m_file);
-                            vramSummary.m_allocationCount = ap.m_allocations.size();
+                            vramSummary.m_allocationCount = static_cast<uint32_t>(ap.m_allocations.size());
                             vramSummary.m_allocatedMemory = ap.m_totalAllocatedMemory;
                             break;
                         }
