+++ Gems/AtomLyIntegration/AtomFont/Code/Source/GlyphBitmap.cpp
@@ -1,6 +1,7 @@
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
@@ -9,10 +10,9 @@
 // Purpose:
 //  - Hold a glyph bitmap and blit it to the main texture
 
-#include <AtomLyIntegration/AtomFont/AtomFont_precompiled.h>
 #include <AtomLyIntegration/AtomFont/GlyphBitmap.h>
 #include <math.h>
-
+#include <CryCommon/Cry_Math.h>
 
 //-------------------------------------------------------------------------------------------------
 AZ::GlyphBitmap::GlyphBitmap()
@@ -118,7 +118,7 @@ int AZ::GlyphBitmap::Blur(AZ::FontSmoothAmount smoothAmount)
                     colorSum += m_buffer[yOffset + x];
                 }
 
-                m_buffer[yOffset + x] = colorSum >> 2;
+                m_buffer[yOffset + x] = static_cast<uint8_t>(colorSum >> 2);
             }
         }
     }
