+++ Gems/EMotionFX/Code/Tests/Vector2ToVector3CompatibilityTests.cpp
@@ -1,7 +1,8 @@
 
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
@@ -24,7 +25,7 @@
 #include <EMotionFX/Source/Parameter/Vector3Parameter.h>
 #include <EMotionFX/Source/Parameter/ValueParameter.h>
 #include <EMotionFX/Source/BlendTreeParameterNode.h>
-#include <MCore/Source/Array.h>
+#include <AzCore/std/containers/vector.h>
 
 namespace EMotionFX
 {
@@ -45,8 +46,8 @@ namespace EMotionFX
             m_blendTree->AddChildNode(finalNode);
             finalNode->AddUnitializedConnection(m_blendNNode, BlendTreeBlendNNode::PORTID_OUTPUT_POSE, BlendTreeFinalNode::PORTID_INPUT_POSE);
 
-            const int motionNodeCount = 3;
-            for (int i = 0; i < motionNodeCount; ++i)
+            const uint16 motionNodeCount = 3;
+            for (uint16 i = 0; i < motionNodeCount; ++i)
             {
                 AnimGraphMotionNode* motionNode = aznew AnimGraphMotionNode();
                 m_blendTree->AddChildNode(motionNode);
