+++ Gems/Maestro/Code/Source/Cinematics/BoolTrack.cpp
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) Contributors to the Open 3D Engine Project. For complete copyright and license terms please see the LICENSE at the root of this distribution.
- * 
+ * Copyright (c) Contributors to the Open 3D Engine Project.
+ * For complete copyright and license terms please see the LICENSE at the root of this distribution.
+ *
  * SPDX-License-Identifier: Apache-2.0 OR MIT
  *
  */
 
 
-#include "Maestro_precompiled.h"
 #include <AzCore/Serialization/SerializeContext.h>
 
 #include "BoolTrack.h"
@@ -38,7 +38,7 @@ void CBoolTrack::GetValue(float time, bool& value)
 
     CheckValid();
 
-    int nkeys = m_keys.size();
+    int nkeys = static_cast<int>(m_keys.size());
     if (nkeys < 1)
     {
         return;
